+++ src/transfers/DepositProvider.ts
@@ -9,9 +9,9 @@ import {
   DepositAssetInfo,
   DepositAssetInfoMap,
   DepositRequest,
-  FetchKycInBrowserParams,
   Field,
   FieldPayload,
+  InteractiveKycNeededResponse,
   KycStatus,
   TransferError,
   TransferResponse,
@@ -86,6 +86,9 @@ function validateEmail(email: string): boolean {
  * serverside/native apps, `getKycUrl`.
  */
 export class DepositProvider extends TransferProvider {
+  public response?: TransferResponse;
+  public request?: DepositRequest;
+
   constructor(transferServer: string, account: string) {
     super(transferServer, account, "deposit");
   }
@@ -97,8 +100,9 @@ export class DepositProvider extends TransferProvider {
    * servers expect)!
    */
   public async startDeposit(params: DepositRequest): Promise<TransferResponse> {
+    const request = { ...params, account: this.account };
     const isAuthRequired = this.getAuthStatus("deposit", params.asset_code);
-    const qs = queryString.stringify({ ...params, account: this.account });
+    const qs = queryString.stringify(request);
 
     const response = await fetch(`${this.transferServer}/deposit?${qs}`, {
       headers: isAuthRequired ? this.getHeaders() : undefined,
@@ -128,6 +132,9 @@ export class DepositProvider extends TransferProvider {
         this.authToken
       }${hash}`;
     }
+
+    this.request = request;
+    this.response = json;
 
     return json;
   }
@@ -209,13 +216,22 @@ export class DepositProvider extends TransferProvider {
    * succeeds, this will be the information needed to complete a deposit. If it
    * fails, it will contain information about the KYC failure from the anchor.
    */
-  public async fetchKycInBrowser(
-    params: FetchKycInBrowserParams<DepositRequest>,
-  ): Promise<KycStatus> {
-    const { response, request, window: windowContext } = params;
+  public async fetchKycInBrowser(windowContext: Window): Promise<KycStatus> {
+    if (!this.response || !this.request) {
+      throw new Error(`Run startDeposit before calling fetchKycInBrowser!`);
+    }
+
+    if (
+      this.response.type !==
+        TransferResponseType.interactive_customer_info_needed ||
+      !this.response.url
+    ) {
+      throw new Error(`KYC Not needed for this deposit!`);
+    }
+
     const kycResult = await fetchKycInBrowser({
-      response,
-      request,
+      response: this.response as InteractiveKycNeededResponse,
+      request: this.request,
       window: windowContext,
     });